refactor: move InfoText out of the Options folder
